FIX Forward autocast_adapter_dtype on all PeftModel subclasses (#3668)
PeftModelForSequenceClassification, PeftModelForTokenClassification and
PeftModelForQuestionAnswering override add_adapter to inject their head into
modules_to_save. All three accept and document autocast_adapter_dtype but
called super().add_adapter() without passing it on, so PeftModel's default of
True always won: adapters were upcast to float32 even when the user explicitly
passed autocast_adapter_dtype=False.

This also affected PeftModel.load_adapter for these task types, since it routes
through the same override; its trailing cast_adapter_dtype call is a no-op when
the flag is False, so the adapter stayed float32 after loading.

Adds a common test, _test_add_adapter_no_autocast_adapter_dtype, parametrized
over the base model dtype (float16 and bfloat16). It is called from
test_seq_classifier.py and the new test_token_classification_qa.py, which cover
the affected task types, and from test_decoder_models.py and
test_encoder_decoder_models.py, which are not affected but guard against a
regression.

+++ src/peft/peft_model.py
@@ -1893,7 +1893,12 @@ class PeftModelForSequenceClassification(PeftModel):
             else:
                 peft_config.modules_to_save.extend(classifier_module_names)
 
-        return super().add_adapter(adapter_name, peft_config, low_cpu_mem_usage=low_cpu_mem_usage)
+        return super().add_adapter(
+            adapter_name,
+            peft_config,
+            low_cpu_mem_usage=low_cpu_mem_usage,
+            autocast_adapter_dtype=autocast_adapter_dtype,
+        )
 
     def forward(
         self,
@@ -2744,7 +2749,12 @@ class PeftModelForTokenClassification(PeftModel):
             else:
                 peft_config.modules_to_save.extend(classifier_module_names)
 
-        return super().add_adapter(adapter_name, peft_config, low_cpu_mem_usage=low_cpu_mem_usage)
+        return super().add_adapter(
+            adapter_name,
+            peft_config,
+            low_cpu_mem_usage=low_cpu_mem_usage,
+            autocast_adapter_dtype=autocast_adapter_dtype,
+        )
 
     def forward(
         self,
@@ -2974,7 +2984,12 @@ class PeftModelForQuestionAnswering(PeftModel):
             else:
                 peft_config.modules_to_save.extend(qa_module_names)
 
-        return super().add_adapter(adapter_name, peft_config, low_cpu_mem_usage=low_cpu_mem_usage)
+        return super().add_adapter(
+            adapter_name,
+            peft_config,
+            low_cpu_mem_usage=low_cpu_mem_usage,
+            autocast_adapter_dtype=autocast_adapter_dtype,
+        )
 
     def forward(
         self,
